fix: a lot

+++ src/components/Settings/DeleteAccount.tsx
@@ -1,7 +1,6 @@
 import { useState } from 'react';
 import { useNavigate } from 'react-router-dom';
 import { useUser } from '../../hooks/useUser';
-import { StyledDeleteAccount } from '../styles/DeleteAccount.styled';
 
 interface Props {}
 
@@ -16,35 +15,36 @@ export const DeleteAccount: React.FC<Props> = () => {
   };
 
   const handleDeleteAccount = () => {
-    deleteAccount();
-    navigate(`../`);
+    let isExecuted = window.confirm('Are you sure to delete your account?');
+    if (isExecuted) {
+      deleteAccount();
+      navigate(`../`);
+    }
   };
   return (
-    <StyledDeleteAccount>
+    <div>
       <div className='tittle'>
         <h2>Delete Account</h2>
-        <div>
-          {isOpen ? (
-            <button className='buttoncancel' onClick={toggleIsOpen}>
-              CANCEL
-            </button>
-          ) : (
-            <button className='buttonchange' onClick={toggleIsOpen}>
-              Open
-            </button>
-          )}
-        </div>
+        {isOpen ? (
+          <button className='buttoncancel' onClick={toggleIsOpen}>
+            CANCEL
+          </button>
+        ) : (
+          <button className='buttonchange' onClick={toggleIsOpen}>
+            Open
+          </button>
+        )}
       </div>
 
       {isOpen && (
         <div className='delete'>
           <p>
-            This action cannot be undone. This will permanently delete yuor
-            account.
+            This action cannot be undone.
+            <p> This will permanently delete yuor account.</p>
           </p>
           <button onClick={handleDeleteAccount}>DELETE</button>
         </div>
       )}
-    </StyledDeleteAccount>
+    </div>
   );
 };
